+++ input/rapid-fire-state.go
