+++ ui/config/options-screen.go
@@ -3,7 +3,7 @@ package config
 import (
 	"fmt"
 	"retro-carnage/engine/geometry"
-	"retro-carnage/engine/input"
+	"retro-carnage/input"
 	"retro-carnage/logging"
 	"retro-carnage/ui/common"
 	"retro-carnage/ui/common/fonts"
