+++ ui/config/select-screen.go
@@ -5,7 +5,7 @@ import (
 	"math"
 	"retro-carnage/engine/characters"
 	"retro-carnage/engine/geometry"
-	"retro-carnage/engine/input"
+	"retro-carnage/input"
 	"retro-carnage/ui/common"
 	"retro-carnage/ui/common/fonts"
