+++ ui/config/video-options-screen.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"retro-carnage/config"
 	"retro-carnage/engine/geometry"
-	"retro-carnage/engine/input"
+	"retro-carnage/input"
 	"retro-carnage/logging"
 	"retro-carnage/ui/common"
 	"retro-carnage/ui/common/fonts"
