+++ ui/enter-name/screen.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"retro-carnage/assets"
 	"retro-carnage/engine/characters"
-	"retro-carnage/engine/input"
+	"retro-carnage/input"
 	"retro-carnage/logging"
 	"retro-carnage/ui/common"
 	"retro-carnage/ui/common/fonts"
