+++ ui/lets-begin/screen.go
@@ -2,7 +2,7 @@ package lets_begin
 
 import (
 	"retro-carnage/assets"
-	"retro-carnage/engine/input"
+	"retro-carnage/input"
 	"retro-carnage/ui/common"
 	"retro-carnage/ui/common/fonts"
