+++ ui/mission/screen.go
@@ -7,7 +7,7 @@ import (
 	"retro-carnage/assets"
 	"retro-carnage/engine"
 	"retro-carnage/engine/geometry"
-	"retro-carnage/engine/input"
+	"retro-carnage/input"
 	"retro-carnage/logging"
 	"retro-carnage/ui/common"
 	"retro-carnage/ui/common/fonts"
