+++ ui/start/screen.go
@@ -5,7 +5,7 @@ package start
 
 import (
 	"retro-carnage/assets"
-	"retro-carnage/engine/input"
+	"retro-carnage/input"
 	"retro-carnage/ui/common"
 	"retro-carnage/ui/common/fonts"
